--- README.rst
@@ -1,15 +1,0 @@
-ABOUT GeoSpacial
-================
-
-Install
-------------------
-
-pipenv shell
-
-pipenv sync
-
-make init-setup
-
-python manage.py import-global gadm3-6
-
-python manage.py runserver
